+++ src/boundary_map/mod.rs
@@ -9,4 +9,5 @@ pub mod linearization;
 
 pub use crate::boundary_periodic as periodic;
 pub use crate::unstable_manifold as manifold;
+pub mod inflation;
 pub mod uncertainty_radius;